+++ cosmic/src/forms/settings/mod.rs
@@ -1,2 +1,3 @@
 pub mod authorized_user;
+pub mod catalog;
 pub mod directory_settings;